+++ src/app/services/CircleDrawer/circle-drawer.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import {WebGLService} from "../WebGL/web-gl.service";
 import {Circle} from "../../../models/circle";
-import {toCanvasCoordinate} from "../../../models/coordinate";
+import {toCanvasCoordinate} from "../../functions/coordinate";
 
 @Injectable({
   providedIn: 'root'
